Introduce storage interface and minio impolementation, add filtering
Similar to MessageQueueChannel ABC we introduce Storage ABC to keep
processed str data with defined parameters.

We also add filtering for documents, note that we filter by chars,
not by bytes, and we opt for precise character count, assuming that
network and storage is available, we download document and then check
characters, we can filter by bytes or assume average utf-8 and estimate
that. It would save bandwidth, compute and time.

This patch also adds more basic Prometheus metrics to monitor the pipeline

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>